move: src/mocks/chat -> src/chat/__mocks__
